refactor: move config interfaces to types and update related imports
- Move AppConfig and I18nConfig interfaces from config files to types/index.d.ts
- Update imports in config files to use the new location
- Reorder navbar menu items and use nativeName in language switcher
